[IR] Correctly handle captured TPs in LocalDeclarationsLowering
…but only on non-JVM backends.
The JVM backend doesn't handle this well for some reason.

This is needed for supporting rich callable references in the Wasm
backend (KT-74392).

Previously, type parameters of local classes weren't remapped,
so when a local class was lifted outside in LocalClassPopupLowering,
it contained references to type parameters that were no longer in
the scope.