KT-73142 Remove SystemGetEnvSource, add option for including system env
Remove SystemGetEnvSource as it's easy to misuse it and accidentally register all System environment variables as configuration cache inputs,
and worse still it won't be visible in the configuration cache report.

Instead, add an option to ProcessLaunchOptions to include the system environment.

Rename `ProcessLaunchOptions.environment` to `ProcessLaunchOptions.customEnvironment` to avoid confusion.
Instead, use `ProcessLaunchOptions.computeEnvironment`.

^KT-73142