JVM: add LanguageFeature for writing annotations to metadata
Write annotations to JVM metadata only in one of three cases:
1) if the language feature is enabled
2) if the declaration is a local delegated property
3) if this is metadata for optional annotations in the module

For local delegated properties, annotations are needed because there is
no other place to store them, and their absence leads to a real problem
in kotlin-reflect (KT-74337). The compiler does not read annotations on
local delegated properties, so it is safe to write them into metadata
unconditionally.

For optional annotations, it is safe to write annotations as well and
have this behavior depend on the metadata version 2.2, because the
compiler can load annotations from both places (the main annotations
field, and the builtins-based extension). The concern with
kotlin-metadata-jvm of version 2.1 modifying this metadata and losing
annotations is not valid here either because metadata for optional
annotations is located in the `.kotlin_module`, and writing module
metadata is not supported in kotlin-metadata-jvm yet (see
`KotlinModuleMetadata.write`).

Test data changes are mostly reverts of the changes from 11ce2cf6,
except that in
`libraries/tools/kotlinp/jvm/testData/AnnotationTargets.kt` the new
language feature is enabled, to test that annotations are written
everywhere correctly.

 #KT-57919