[IR] Set the private visibility for lifted local classes
In the next commit we will fix the handling of type parameters when
lifting local classes.

However, in order for
`org.jetbrains.kotlin.ir.types.extractTypeParameters` to not collect
the type parameters from outer scopes for an already lifted local class,
we have to set the visibility of those classes to something non-local.

However, sometimes we still need to detect such classes even after
they were already lowered, so we introduce an IR attributes to mark
such classes during `LocalDeclarationsLowering`.