[K/N][tests] Fix testBlockToFunction with gcType=NOOP
The test `testBlockToFunction` was expanded in 48e5add. Part of the test
relies on GC to collect the garbage. This is a natural expectation, but
it doesn't hold when gcType=NOOP. Which made the test fail in this mode.

This commit fixes this by extracting that part of the test to a separate
test that is ignored when gcType=NOOP.