[FIR] Make `FirWhenSubjectExpression` into a property access expression
Deleting `FirWhenSubjectExpression` altogether seems inconvenient:
we may still need to tell apart `$subj$ is Type` from `subject is Type`,
and having a separate node is quite handy. What we really want is to
simplify the DFA logic related to `when` subjects, which we can do by
offloading it to regular property accesses.

The change in `coneDiagnosticToFirDiagnostic.kt` is needed to prevent
"strange" errors on short conditions like `is Type ->` or `value ->`,
like `<!UNRESOLVED_REFERENCE!>""<!>`. See `guardWithNoSubject.kt`.

The change in `VariableInitializationCheckProcessor` is required to
prevent `<!UNINITIALIZED_VARIABLE!>is Type<!> ->`. See
`unsupportedVariableDeclarationsInWhenSubject.kt`.

^KT-75704 Fixed