Tests: refactor FIR metadata loading test suppression
Always compare test dump, and then suppress a failure, if any, in
FirMetadataLoadingTestSuppressor.

Previously, if the test had an incorrect dump and was muted by a
directive, we did not compare the dump, so the test was successful, and
FirMetadataLoadingTestSuppressor failed with the "looks like this test
can be unmuted".

Also, extract the directive that is used to suppress, into the abstract
test class, to avoid computing it afterwards based on external factors
(e.g. frontend kind).