[KGP] Experimental: support incremental changes in inlined local classes
In scenarios where a local class declaration is inlined, Kotlin IC might
fail to detect the change in inline function, and skip recompiling some
of its call sites.

To fix that, we implement a brute force approach. INSTANCE usages in
inline methods are extracted from the JVM bytecode. Then inline function
"state" hashes are calculated as a multi-hash of function's own hash and
used local classes' hashes.

This behavior is disabled by default, because it is in testing.

Gradle option
kotlin.internal.classpathSnapshot.parseInlinedLocalClasses=true

would enable the new logic.

^KT-62555 Fixed
^KT-75529