[CODEOWNERS] Expand ownership of :compiler:tests-compiler-utils & :compiler:tests-integration
The modules in question now contain a lot of code that is implemented, used, and maintained for varying purposes by different compiler teams. A requirement of change approval by Kotlin Frontend for all of this code leads to formalities that don't really contribute to overall code quality. Given that the modules have no clear separation of code owned by different teams, it makes sense simply to share ownership with all teams involved.