Update inlineFuncWithExposedComposableSingleton.txt after allowing indy
The described change in ABI in this txt file is accepted as OK since
the leakage happend through the following method:

```
    public final method getLambda-1$main(): kotlin.jvm.functions.Function2
```

Additions described in this change are OK as the ComposableSingletons
are no longer leaking.

Removals described in this change are OK as they are not used
by older consumers, therefore not breaking their support.

^KT-75768 fixed