[K/N] Fix local lifetime support in EA, allow disabling it
If escape analysis infers an object lifetime to be `local` (e.g., when
the object escapes a loop iteration but not the enclosing function),
then it mistakenly resets it to `global`.

Considering that propagation of escaping for "exiled objects" is now
disabled, this can make the heap object refer to a stack object.
Which is wrong for CMS because it can scan this reference after the
stack object has become invalid.

This commit fixes this by not resetting inferred `local` lifetime to
`global` anymore.

It also adds a new binary option, `enableLocalObjects`, to allow
disabling the local lifetime support.

^KT-75317 Fixed