[PowerAssert] Introduce HiddenNode instead of using ConstantNode
Instead of using ConstantNode for expressions that do not need to be
displayed, introduce a new node type HiddenNode. This node type can be
excluded from the diagram but still extracted into a temporary variable.
This makes sure the program execution order is still correct even when a
value is not displayed.

This would also allow a parent value to not be displayed while a child
value is displayed. For example, a function call could be excluded while
including the arguments of the function call.

^KT-75263