[K/N][runtime] Wrap MarkQueue into the new type MarkState in mark utils
This is a preparation for KT-75861.

Fixing the issue will require `processFieldInMark` to enqueue local
objects to a separate queue.

This commit introduces a new type, `MarkState`, currently holding only
the `MarkQueue` and makes all mark utilities (including
`processFieldInMark`) take `MarkState` argument instead of `MarkQueue`.
This way, adding a new queue to `MarkState` will be enough for making it
available in `processFieldInMark`.