[IR Dump] New option to render types & annotations with unbound class symbols
With the new option, if an unbound symbol is met in `IrSimpleType`,
the IR dumper should try to guess the fully qualified name of the class
by the symbol's signature before rendering the standard "unbound symbol"
text.

Here is an example. With this option, the following part of the dump
of deserialized IR
  TYPE_OP[939, 944] type=<unbound IrClassSymbolImpl><kotlin.Any?> origin=IMPLICIT_CAST typeOperand=<unbound IrClassSymbolImpl><kotlin.Any?>
should be rendered as
  TYPE_OP[939, 944] type=kotlin.coroutines.SuspendFunction0<kotlin.Any?> origin=IMPLICIT_CAST typeOperand=kotlin.coroutines.SuspendFunction0<kotlin.Any?>

^KT-75986 Fixed