[IR Dump] New option to suppress real member access expression arg names
Using this option, it is possible to render just indices of type &
value arguments for `IrMemberAccessExpression`s instead of their names.

This is helpful for IR deserialization tests running with the IR inliner
enabled. In such tests, the IR before serialization may have member
access expressions referring to unbound function symbols. For such
expressions it's not possible to get a function to collect real
parameter names. Thus, the expression is rendered with just indices:
  CALL[1159, 1181] ...
    TYPE_ARG 1: kotlin.Array<kotlin.Any?>
    ARG 1: TYPE_OP[1150, 1158] type=kotlin.Array<out kotlin.Any?> ...

However, after the deserialization, the function symbol in
the expression becomes bound, and it's possible to collect real names:
  CALL[1159, 1181] ...
    TYPE_ARG T: kotlin.Array<kotlin.Any?>
    ARG <this>: TYPE_OP[1150, 1158] type=kotlin.Array<out kotlin.Any?> ...

This causes test failures.