[IC] Always add files of expect declaration with lenient stubs to dirty set
This is necessary to support incremental compilation with lenient mode.
In an incremental compilation, a real actual declaration can be added
for an expect declaration which previously had a lenient stub.
In this case, the file with the expect declaration must be part of the
compilation, otherwise we get an ACTUAL_WITHOUT_EXPECT error.
Also, we clean the class file of the stub so that we don't end up
with two class files for the actual declaration.

#KT-76041 Fixed