[Build] Make the `install` task transitive
This also changes the scope of some dependencies in some modules from `api` or `implementation` to `compileOnly`. That is because those modules are published, but those dependencies are not. Such configuration effectively publishes artifacts that one cannot consume. After changing the scope, it still remains problematic, e.g. consumers have to declare additional runtime dependencies to make it work, but this is a separate problem to be solved via KT-77274.
^KTI-2200 Fixed