Add rest libraries symbol provider to binary lib session as fallback
The new project structure provider (KTIJ-31422) dropped the binary
library module dependencies. As a result, we cannot find a symbol
with a class ID from the binary library session. In the case of source
library session, KT compiler adds the rest libraries symbol provider as
a fallback, but it does not add the fallback for a binary library
session. As a result, we have a real-world FIR compiler facility CodeGen
exception: KTIJ-33530. This commit adds a rest libraries symbol provider
to binary lib session as fallback to fix the CodeGen exception.

^KTIJ-33530