[FIR] Make sure FirBlock has the correct result type from lambda
When a lambda is the last expression in a FirBlock, the resolved type
is usually not known yet due to resolution order. To delay the setting
of the result type, previously a cone error was set and then reset
after call completion. This was removed a while ago and replaced with
simply a unit type, thus causing some FirBlocks to have an incorrect
resolved type when compared with their parent or the last expression of
the block.

Revert to using this temporary cone error and document it better so it
doesn't get removed without a correct replacement.

^KT-75712 Fixed