[K/N] Worked around a lldb bug
There is a bug in lldb around evaluation & stepping into an inlined function:
the current frame gets dropped which leads to loss of the current source location.
This commit works around this by saving the frame information before calling the evaluation
and restores it after the evaluation has been finished.

 #KT-75991 Fixed