[LL] FileStructureElement: improve logic for classes
This commit unifies the logic around a class file structure element.
Now all such conditions are the same across different visitors.
Also, this commit makes the visitor logic resistance to changes
around `FirRegularClass` as now it properly handles member declarations
and explicitly separates them from other class parts such as annotations
or an implicit constructor.

Another noticeable change is that `FirDanglingModifierList` is not
a part of the class file structure element. It was separated in
da027f3f062f60766ac4cc933317d249315a4d56 commit, but redundant
parts are left in the element

^KT-76331 Fixed