Fix non-transitive implementation of AccessorSignature.equals()
The implementation of `equals()` in `AccessorSignature` allows
comparing an instance of `AccessorSignature` to both
- another instance of `AccessorSignature`
- and an instance of some other signature class that may represent
  a pure accessor (without the property).

However, there are no other signature classes that allow comparing
an instance of self to another signature class. Which means that
the implementation of `equals()` in `AccessorSignature` does not
have a property of transitiveness and is probably incorrect.

With this change `AccessorSignature.equals()` should allow comparing
only to other instances of `AccessorSignature`.