Reflection: use kotlin-metadata-jvm for KClass.nestedClasses
The change in `loadClass` is needed because otherwise
`companionObjectInstance` would start to work incorrectly for built-in
classes (e.g. `String`). For those classes, we always need to take
stdlib's (or reflect's) class loader. This roughly corresponds to the
`else` branch in the `ClassDescriptor.toJavaClass` utility, which was
used before this change.

 #KT-75505