[FIR] Unit literals as when branches should be considered used
It is a common pattern to use a Unit literal instead of an empty block
for a when branch that shouldn't have any behavior. If the when is
unused, this marks the Unit literal as unused as well. But really the
Unit is being used for a different purpose and should not be reported.

^KT-76290 Fixed