[AA] Add symbol tests for function inside dangling annotation
Note: the testdata is committed for PSI-based results, so currently
there is an error "Non-PSI version doesn't equal to the PSI-based variation".

The bug in the PSI-based version is that
the symbol locations should be LOCAL instead of TOP_LEVEL / CLASS.

Also, the `firSymbol`-based version returns different results:
the location is LOCAL (which is correct), but the `callableId` is
not null (should be null for local declarations).

Both of these problems will be fixed in the following commit.

KT-75534