[FIR2IR] If necessary, cast when subject initializer with intersection type
When subject variables have special rules.
When the initializer is smart-casted, in FIR the variable has the
type of the original expression but in IR, it must have the
smart-casted type.
That's why we don't set the type of the variable from the FIR
property.
But when the initializer has a weird intersection type, that is
approximated to something that isn't a subtype of the variable
type (e.g. Any), we end up with a variable with the wrong type.
For now, we insert a cast when such a situation arrises.
Hopefully, it will be cleaned up in #KT-63348.

#KT-75983 Fixed