[FIR] Getters/setters with annotations can still be considered default
When serializing an FIR property accessor, only the presence of an
annotation on the setter parameter should indicate the setter is
non-default. Previously, any annotation on the accessor would cause it
to be marked as non-default, which caused an inconsistency during
incremental compilation and friend module compilation when compared with
a full compilation.

^KT-69075 Fixed