Partial revert "Reduce number of I/O operations for libraries with many files"
Unfortunately, the mentioned commit caused a significant performance
degradation (1% on Kotlin project analysis, 2% on intellij project analysis).
This is probably provoked by the fact that the older version
(to which we are returning now) didn't call .toRealPath() in all when
branches (see LibraryList.accepts), but the newer one does it always.

This reverts (partially) commit fb30915635635c74ee1b490cabb06ff28cc69256.

#KT-76395 Fixed