Refactor Compose function lowerings to the new parameter API
The old API is going to be removed in 2.2.0 and this migration ensures that we don't use any of the old value parameter APIs.

This change is the first in series of commits that migrate to the new API. This change migrates ComposableFunctionBodyTransformer and ComposableDefaultParamLowering. It also includes a clean up for calculations of changed parameters that relied on deprecated API.

Test: Compiler tests
^KT-73367 Fixed