[K/N][runtime] Implement a workaround for KT-75992, disabled by default
This commit introduces a new binary option:
  coreSymbolicationImageListType=only_kotlin

When enabled, it makes K/N runtime create a CoreSymbolication
symbolicator by supplying only the Kotlin binary image to it, instead of
all loaded binary images. This way, the execution doesn't rely on
`_dyld_process_info_create` which is affected by a bug in dyld.

See the detailed explanation in the issue and in the source code
comment.

^KT-75992