[Inline] Remove private inlined symbols from the public scope
A private symbol should not be accessible from a public scope.
Partial linkage will correctly fail with an error if something
like this appears. In this commit we remove private symbols
that were inline in a public scope but keep symbols in place if they
were inlined in a private scope.

#KT-76181 Fixed