[FIR] Check for uninitialized receiver of implicit invoke operator
Unfortunately, there is no QualifiedAccessNode for an implicit receiver
of an invoke operator. This has to do with the order in which arguments
and the receiver are resolved, and the CFG being unable to correctly
order nodes for this situation. So, check each FunctionCallEnterNode
instead, and if it is for an implicit invoke call, check the receiver
at the FunctionCallArgumentsExitNode.

^KT-74326 Fixed