[K/N][runtime] Fix `kCSNow` value in interaction with CoreSymbolication
Apparently, we got the value of `kCSNow` constant from CoreSymbolication
wrong.
The right value is `(1ull<<63)`, which is revealed in Swift backtracing
implementation:
https://github.com/swiftlang/swift/blob/1745467062bfca5d686a979cf73354ffc67e3257/stdlib/public/RuntimeModule/modules/OS/Darwin.h#L241C28-L241C34

Kotlin/Native runtime used `(1ull<<63) - 1` instead, which is actually
`kCSEndOfTime`, according to the same source.

The difference wasn't important when creating the symbolicator with
`CSSymbolicatorCreateWithPid(getpid())`, but it affects
`CSSymbolicatorCreateWithBinaryImageList` which will be used soon.

This commit fixes this by using the correct value for `kCSNow`.

^KT-75992