KLIB Loading: Introduce KLIB loader - the replacement of KLIB resolver
The `KlibLoader` component helps to load `KotlinLibrary`s based on
the supplied paths.

This component also performs a limited set of basic checks, such as
- Checking a platform and a specific target
- Checking ABI version compatibility.

Any other checks, if they are necessary, should be performed outside
`KlibLoader`.

Sorting of libraries, if it's really necessary, should also be
performed outside `KlibLoader`.

`KlibLoader` does not add to the `KlibLoaderResult` any transitive
dependencies that any of the loaded libraries might have. Only
the libraries explicitly supplied to `KlibLoader` may appear in
`KlibLoaderResult`.

^KT-65837