K2: Fix inference failure on delegated implicitly typed var inside PCLA
To repeat K1 behavior, we don't analyze `setValue` for implicitly typed
delegated _var_ withing delegation inference session and only run after
the completion.
(see usages of `isImplicitTypedProperty` at
 FirDeclarationsResolveTransformer.transformPropertyAccessorsWithDelegate)

But before this commit when completion writing applies to the PCLA
lambda, it recursively tries to go inside all containing declarations,
including the property and its accessor which isn't resolved yet. Thus
failing with "Expected expression 'FirSingleExpressionBlock' to be resolved".

The idea of the fix is just avoiding visiting any declarations
that here not only remove redundant workload but also resolves the issue.

This commit doesn't fully fix KT-76171 because the case with anonymous
objects is only covered in the next one.

^KT-76171 In Progress