[Analysis API] Return more deprecated methods needed for Android Lint
Currently, the Compose Lint checks bundled with the Compose library are
built against the old version of the Analysis API. The reason is that
these checks need to be compatible with a wide range of Android Studio
and AGP releases, and Android Lint does not allow to have check forks
specific to certain IJ platform releases.

The previous set of methods turned to be not sufficient to make the
checks working.

The long-term plan is to eventually migrate the checks to the actual
version of the Analysis API.

^KT-75897 Fixed