[FIR] Java: support decompiled light classes as source for annotation arguments mapping
This is the follow-up fix for KT-76540.
The problem is that compiled Kotlin classes are wrapped into the
decompiled light classes in the IntelliJ Kotlin plugin.
As a result, `JavaClassImpl#assertNotLightClass` expectedly
fails each time the mapping accesses `methods` method.
The issue is covered by the plugin tests as the current Kotlin repo
infrastructure is not capable of reproducing the issue.

^KT-76540
^KT-76625 Fixed

(cherry picked from commit a7ab37413623a1cb5b9003d1aa5c3785e1221910)