[FIR] Support negative information for not very trustworthy classes
We've always allowed our exhaustiveness checks to consider them "safe",
so best we can do is to maybe show a warning when the user checks for
his own classes with potentially problematic `equals()`.

One nuance, though, is that we trust `equals()` of all objects, regardless
of whether they are `data` or not. The default `equals()` of non-`data`
objects is actually unsafe for exhaustiveness (if we get a second
instance of `MyObject` at runtime, it will return `false`), but because
"just adding" `data` is not always possible (e.g. when the user wants
to inherit a specific version of `toString()` from the supertype), we
choose to silently ignore it.

Additionally, notice that, unfortunately, scenarios where
`SAFE_FOR_SMART_CAST` and `TRUSTED_FOR_EXHAUSTIVENESS` act as unrelated
values are closer than we thought :( This is probably a sign that we
should to replace the enum with a pair of two properties, but because
we consider `object`s safe, we can get away with a simple enum.

^KT-76635