PL Tests: Introduce two new tests for inline functions
Both tests are supposed to have the same test data. The only difference
is that `inlineFunctionTransformations-no-inliner` should always run
with disabled IR inliner at the 1st compilation phase and
`inlineFunctionTransformations-with-inliner` should always run with
enabled IR inliner. This way we could be able to test ABI compatibility
both without and with inliner.