[IR] Temporarily suppress deprecated old parameter API usages
Our code is almost entirely migrated to tne new API at this point,
we have just a few places left that are expected to be finished in a
matter of days. Here we OptIn to @DeprecatedForRemovalCompilerApi in
those places so that we can mark the old API as with it for before
kotlin 2.2, in the next commit.