[SLC] fix contract violation via super types
Super type mapping triggered redundant type parameters bounds computation
which result wasn't used as the Analysis API side is interested only
in the type argument itself and not in its bounds.
As a side effect of the restriction, type nullability will not be
inferred for type arguments in super types position.

^KT-75689 Fixed