[Tests] JS. Fix CLI tests that fail because of unresolved symbols
Earlier, these tests failed with an error that some library has
not been resolved because there is no library at the given path.

But now this is no more a kind of problem that stops compilation
immediately. Instead, the compilation proceeds and the compiler
fails later with the error about unresolved symbols.

^KT-65837