[Gradle][test] Filter "Invalid connection: com.apple.coresymbolicationd"
When symbolicating K/N stack traces on simulators, this sometimes prints

    Invalid connection: com.apple.coresymbolicationd

Previously, this was printed every time. But after KT-75992 (80518f5)
it is not printed sometimes.

Specifically, `GeneralNativeIT.testNativeTests` runs the same binary
twice, and this line is usually printed on the first run but not on the
second run.
But the test expected this line to be present on the second run as well,
so it fails now.

This commit fixes that by removing the line from the expected output and
from the actual output (if present). By doing so, it effectively
reintroduces an earlier fix (219d5e3) which was removed in a
refactoring.

^KT-76748

(cherry picked from commit 0b01e8af78291a2e30edecbc955525b35c5dcb95)