JS & Wasm IC: Use CompilerConfiguration to read library paths
Use the `CompilerConfiguration` instance to read library paths:
- all libraries: `configuration.libraries`
- friends: `configuration.friendLibraries`
- included library: `configuration.includes`

This is a preparatory step to make it possible to use the same
KLIB Loader entry points in the JS/Wasm compiler and in JS/Wasm IC
engine.

Also, it does not make sense to pass library paths separately as
`CacheUpdater` anyway requires a `CompilerConfiguration` instance.

^KT-65837