[Docs] Mark new kotlin.time types as new in 2.1.20
The KotlinLang Docs for e.g.
[`Instant`](https://kotlinlang.org/api/core/kotlin-stdlib/kotlin.time/-instant/)
specify that Instant is available since Kotlin 2.1.
However, it is only available from 2.1.20. The same applies for `Clock`.

When attempting to use these types in a Kotlin 2.1.10 project,
I was left confused at the inability to import them.
I eventually found the reason in the Kotlin 2.1.20 release changelog.

SinceKotlin's version field supports patch-level versions:
> `@property version The version in the following formats: `<major>.<minor>`
> or `<major>.<minor>.<patch>`, where major, minor and patch
> are non-negative integer numbers without leading zeros.

This PR includes the patch version for the Instant, Clock,
and the relevant Instant top-level functions.