[JPS] Replace the Maven serialization plugin with the regular compiler plugin
IDEA downloads Kotlin dist via the `kotlin-dist-for-jps-meta` artifact and creates a layout similar to Kotlin dist. This logic has special treatment for kotlin-serialization-compiler-plugin for JPS due to its renaming in the past. This special treatment saves an artifact named as `kotlin-maven-serialization*` as `kotlinx-serialization-compiler-plugin.jar` to save it in the dist. A proper long-term solution should be provided in IDEA via KTIJ-33877. This new artifact should be removed in future and is not intended for any usage except this workaround.
This cannot be tested in the Kotlin repository as here we have the real Kotlin dist that contains proper artifacts.
^KT-76495 Verification Pending