[FIR] Account for `UNSAFE_EXHAUSTIVENESS` in the source code
`CompressedPointsToGraph.NodeKind.Return` defines a custom equals and
the compiler doesn't trust it. Even though it's identical to the default
one.

`KonanTarget` probably deserves some refactoring turning the `object`s
into `data`, but this code seems to still be using an old compiler that
emits:

> The feature "data objects" is only available since language version 1.9

The diagnostic suppression is needed for the bootstrap. Using `data`
`object`s and getting rid of the manual `equals()` implementation makes
sense as no one is going to create other instances of `KonanTarget`,
different from the ones listed.