[K/N][tests] Move codegen/box tests to a separate module
Currently, the codegen/box tests runners for Kotlin/Native are placed
in the `:native:native.tests` project, among with other tests.

This leads to poor DX:
* The project has many test tasks, so running tests in IDE is confusing:
  running a test through a "gutter" asks the user to select a test task.
* The "main" test task, "test", requires "platform libraries",
  so running a test with this task takes a lot of time, and platform
  libraries are rebuilt after any change in the compiler.

To overcome the latter, one could select the `codegenBoxK2Test` task
when asked. But that's obscure.

This commit improves this by moving the codegen/box test runners for K/N
to a separate module, `:native:native.tests:codegen-box`.

So now running a K/N codegen/box test in the IDE through a gutter
doesn't ask to select a test task and doesn't build platform libraries.

^KT-72597 Fixed