[K/N][tests] Don't create temp files in the project root in cli-tests
The tests from `:native:native.tests:cli-tests` create temporary files
in the project root directory.
The reason: the test engine runs the compiler with the working directory
set to the project root, and the tests don't specify the output
directory. So the compiler emits output to the project root.

This commit fixes this by setting the output directory to `$TEMP_DIR$`
in all test data for these tests. This is aligned with CLI tests for
other backends.

^KT-75311 Fixed