[Build] Don't use `backend.jvm` as an api dependency of `cli-base`
Currently, `:compiler:cli:cli-base` defines an `api` dependency on
`:compiler:backend.jvm`. So the latter becomes visible in many modules
using the former, making it easy to accidentally use JVM-specific code
in other backends.

`cli-base` does use some code from `backend.jvm`, so the dependency is
necessary. But it can be made an `implementation` dependency so that
other modules can't see `backend.jvm` because of this.

This commit does exactly that.

^KT-73707 Fixed