[K/N] Remove deprecated kotlin-native/Interop/JSInterop
This component was used for the "JS interop" capability of the wasm32
target in Kotlin/Native.
The target and its capability are long gone. The API of the component
was error-deprecated in 3ec4f2e.

This commit removes the component and its traces.

It has nothing to do with the modern replacement of that target,
Kotlin/Wasm.

^KT-61549 Fixed