[Test] Fix JSpecify CLI tests to reference the correct annotation dir
The JSpecify annotations in the code were only being recognized because
they were fully-qualified. This is why none of the strict cases were
reporting an error for the `@Nullable` case: because the annotations
weren't resolved to see they target `TYPE_USE` and not `FUNCTION`. This
meant the return type in FIR had the incorrect annotations.

^KT-53836