[Maven] Avoid using transitive dependencies of `maven-core`
`maven-core` is not officially stated to be available at runtime (though, it appears to be, but without some transitive dependencies). Perhaps, we should not use it at all, but we have usages of symbols available only there and this logic is covered by tests. However, we used transitive dependency of it (Guava) that by lucky coincidence was available in kotlin-compiler artifact. After migration to BTA (KT-73012), this dependency is not anymore there, so build running the logic that uses Guava, started to fail.
This change migrates the problematic logic to use relocated Guava from the Kotlin compiler and removes transitive dependencies of `maven-core` from the `provided` scope, replacing it by dependency on `maven-plugin-api`.
^KT-77036 Verification Pending