[K/N][llvm][build] Change -DCMAKE_OSX_DEPLOYMENT_TARGET=11.2 to 11.0
This commit makes llvm_builder pass -DCMAKE_OSX_DEPLOYMENT_TARGET=11.0
to the LLVM build instead of 11.2. So, it adjusts the macOS deployment
target version for the built LLVM.

This is the least intrusive way to fix the annoying K/N build warnings
like

     ld: warning: object file (...) was built for newer 'macOS' version (11.2) than being linked (11.0)

The reason behind these warnings: LLVM is built for macOS 11.2, while
K/N compiler dylibs are linked for macOS 11.0.
The first version comes from the LLVM CMake variable described above.
The second version is the current minimal macOS version supported by
K/N (defined in konan.properties).

So, while bumping the latter is possible, it is a user-visible change,
and adjusting the former is not.