FIR: Fix scope for annotation argument resolution on a nested class
The problem is that we see member scope there, thus trying to resolve
members as arguments, though the class is nested.

The idea of the fix is that we should use special tower mode
for the whole class header
(it's already done for class bodies, see the second call-site of
`withSwitchedTowerDataModeForStaticNestedClass`).

^KT-76357 Fixed
^KT-77137 Fixed