[Analysis API FIR] KaFirStopWorldCacheCleaner: mitigate deadlock
The commit fixes the deadlock, but the solution still is not perfect
as there is a theoretical chance of a livelock.
See the example:
```kotlin
fun performSomeAction(declaration: KtDeclaration) { // parent thread
    analyze(declaration) { // 1 parent `analyze`
        val newThread = thread { // child thread
            // `scheduleCleanup` happens here

            analyze(declaration) { // 2 child `analyze`

            }
        }

        while (newThread.isAlive) { // 3 parent waits for child
            ProgressManager.checkCanceled()
            newThread.join(50L)
        }
    }
}
```

In the example above, the child thread will reach the loop inside
`waitForCleanupIfNeeded` while the parent thread will reach the loop (3)
to wait for the child thread to finish.
Now imagine that it is so coincident that the parent thread is always in
the `Thread.State.RUNNABLE` state when the child makes another iteration
to check for its state via `threadsWithAnalyze`.

Here we have the livelock.
Hense, the more actions similar to `performSomeAction` we run in parallel,
the higher the probability of ending up in the livelock

^KT-76577 Fixed
^KT-77093