[JS] Delete a JS test that doesn't make sense
If we recreate the same module structure using Gradle and try to
run it, we will get a different result
("OFail AT" instead of "OK").
This is because the result of this test heavily depends on
how our test infrastructure invokes the compiler
(namely, its second stage), which it does quite differently
from what Gradle does on real-world projects.

As we're trying to refactor the JS test infra to make it closer
to real world compiler invocations (KT-77027),
we have to get rid of this test.