[Serialization] Use `KtElement` instead of `PsiElement` for diagnostics
`PsiElement` gets shaded in the embeddable version of the compiler, which
might cause some incompatibilities when using embeddable compiler with
non-embeddable plugin and vice versa. After some of the previous commits,
such a problem occurred with REPL tests.

Theoretically, it also could be solved by fixing the dependencies in
these tests, but this change is useful anyway, as it reduces the
difference between shaded and not shaded version of the plugin, which is
always good.