[IR] Don't allow non-inlined IrFunctionReferences
We migrated to IrRichFunctionReference in all klib backends, and unlike
IrFunctionReferences, IrRichFunctionReference does support inlining.
It means we can expect all function references to be inlined.

Note: inlining a function reference means that the body of the
referenced function is inlined into the body of KFunction.invoke(),
however the call to invoke() itself, if it occurs, is not inlined by
FunctionInlining.
In some situations like `(::foo)()`, it might be inlined later on, as
another optimization.

#KT-69457 Fixed