[test] Prepare a JS test for migrating to the new CLI infra (KT-77027)
With the new CLI infra this test would fail with `UNRESOLVED_REFERENCE`
on `assertEquals` calls in the `main` module.

This is because the `assertEquals` function comes from
`js/js.translator/testData/_commonFiles/asserts.kt`, which is added
to each test by `JsAdditionalSourceProvider`.

Specifically, `JsAdditionalSourceProvider` adds such common files
to modules that have no dependencies (to prevent duplicate symbol
errors during linkage).
In this case, it was `ModuleA1`, on which `main` depends only
transitively, which is why the frontend wouldn't see `assertEquals`
from the `main` module.
(It is unclear why it worked before the migration.)

We fix this by moving the `// NO_COMMON_FILES` directive from `ModuleB`
to `ModuleA1`, thus forcing those common files to be added to `ModuleB`,
on which `main` depends explicitly and we're good.