Reflection: minimize usages of descriptor renderer
Only use it to render types, which will be removed as well in a
subsequent commit.

Receiver parameters for callables are extracted into a separate field
`KCallableImpl.receiverParameters` because in case of a bound callable
reference, there's no way to extract the bound receiver type from the
public `KCallable` interface.

 #KT-75463