[IR] Unify SharedVariablesManager across non-JVM backends
Introduce a new class in the standard library for creating boxes
for shared variables.

In JS, we lower calls to these functions into operations with object
literals.

In Wasm and Native, they are not handled specially in any way.

This is a kind of regression for Wasm, since there used to be box
classes for primitive types there.
This is fixed in the next commit.

KT-70295