[K/N] Remove handling of "removed targets" from `KonanDriver`
`KonanDriver` had special support for "removed targets": whenever a user
of the command-line compiler requested compilation for such a target,
`KonanDriver` made the compiler emit an understandable error about the
status of this target.

The targets were removed quite some time ago, so it is safe to assume
that all interested parties have already acknowledged the change and no
longer need clarifications.

Therefore, this commit removes this special handling.

^KT-59008