KT-77210 [LL] Avoid building designation for properties from primary constructors in `ContextCollector`
If a property from primary constructor is targeted, the designation path
before this change looked like this: `File -> Class -> Property`

When `ContextCollector` follows this path, it builds
the `towerDataContext` incorrectly by skipping the scope setup
for the constructor parameters context (which happens
by calling `BodyResolveContext.forConstructorParameters`).

In the primary constructor, there should be no implicit receiver of
the outer class' type, because it becomes accessible only later in
the class body.

After this change, the designation will target the outer class'
constructor instead: `File -> Class -> Constructor`.

This might cause different parts of the class to be resolved,
but overall it should not trigger the resolution of the whole file
(this concern was mentioned in the bed73fd6 when the
removed condition was added).
So performance-wise this change should be fine.

This change also fixes the following issues in Kotlin:
- KTIJ-33989 - auto-import will no longer propose extension functions
in primary constructors
- KTIJ-29250 - code completion will no longer propose members and
extensions in primary constructors
- KTIJ-34024 - code completion will start seeing parameters declared
before properties in primary constructors

^KT-77210 Fixed
^KTIJ-33989 Fixed
^KTIJ-29250 Fixed
^KTIJ-34024 Fixed