[stubs] KotlinPropertyStub: add `hasBackingField` flag for binary properties
This flag is necessary to use it during stub-based fir deserialization.
Technically, it is possible to extract `hasBackingField` information for
decompiled stubs. However, it will be used only during library stubs
deserialization, so there is no need to complicate the logic as it is
not so simple to propagate the flag via text

^KT-74777