[Gradle] Add CheckKotlinMultiplatformDependenciesTask
It reports readable diagnostic when dependency  doesnt support required
target. This can happen when users add a new dependency without
knowing that it may not support all targets that user's project has
enabled. Gradle by default reports quite cryptic and extensive
error message with explaining that some variants can't be found.
However user should not know anything about Gradle variants, they just
declared a dependency. It is our job to tell them that it is not
good enough.

These changes are doing it heuristically and
only for external dependencies for project dependencies
it is impossible to intercept resolution errors.

^KT-69790 Verification Pending