[CLI] Determine test/production kind of app env depending on the configuration key
This is needed to create the app environment in the test mode for
CLI-based tests. The compiler actually doesn't really care for this mode,
but it's important for AnalysisApi tests, as prod and test mods defines
different rules regarding read/write actions. So if there is a module
with both compiler CLI-based tests, we should always create the app env
in Test mode (even if CLI-based tests execute first). Otherwise AA tests
will fail with read-write access not allowed exceptions.

The proper cleanup of app creation utilities would be addressed
within KT-77328.