[FIR2IR] Refactor implicit cast insertion
The visitor implementation in Fir2IrImplicitCastInserter is removed.
Instead, the implicit cast insertion is pulled into the callsites.

For IrStatementContainers (mainly IrBlock, IrComposite and IrBlockBody),
we call coerceStatementsToUnit explicitly where necessary.

visitSmartCastExpression is replaced with handleSmartCastExpression.

visitWhenExpression is replaced with handleWhenExpression.

Try, catch, finally and return elements are handled at the creation site
in Fir2IrVisitor.

As a result, we can remove the call to accept(implicitCastInserter, it)
in convertToIrExpression altogether.

#KT-63348