[FIR] Check missing supertypes of the actual dispatch receiver for qualified access
It's more correct to take the type of the dispatch receiver expression
from the call-site, as the dispatch receiver might be less specific
due to lack of fake-overrides. But at the backend the f-o from the
more specific type will be used, so if one of supertypes is missing then
it could cause unexpected errors.

^KT-73611