[FIR2IR] Remove redundant unit coercion in whens
Remove handleWhenExpression because it didn't do anything.
In the `if (irExpression is IrBlock)` case, the block only
ever contains a subject variable declaration (doesn't need coercion)
and the when expression (last expression).
Note that we didn't even process the when itself in this case and it
didn't matter (see reason below).
In the other case, we would process the when branches but in fact they
were already processed in Fir2IrVisitor.toIrWhenBranch were we call
prepareExpressionForGivenExpectedType.

#KT-63348