Metadata: remove multi-field value class representation
These fields are a part of an experimental feature whose fate is unknown
at this point. These fields are not exposed in kotlin-metadata-jvm,
which presents a challenge when rewriting kotlin-reflect (KT-75463)
because MFVC are supported in kotlin-reflect and there are many tests on
it, but currently it uses deserialized descriptors, where representation
is loaded from metadata directly.

It seems that we can load MFVC representation from the primary
constructor parameters instead. It won't work in case the constructor is
private and metadata has been modified (e.g. all private declarations
were removed), but it's good enough for an experimental feature for now.

 #KT-77183 Fixed